pyproject.toml
[tool.poetry]
name = "zangalewa"
version = "0.1.0"
description = "An AI-powered command-line assistant for bioinformatics and technical workflows"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/yourusername/zangalewa"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.0.0"
textual = "^0.20.0"
aiohttp = "^3.8.3"
requests = "^2.28.2"
sqlite3 = "^3.41.2"
faiss-cpu = "^1.7.4"
psutil = "^5.9.4"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
openai = "^0.27.0"
anthropic = "^0.2.9"
pytest = {version = "^7.3.1", optional = true}
hypothesis = {version = "^6.70.0", optional = true}
sphinx = {version = "^6.2.0", optional = true}
mkdocs = {version = "^1.4.2", optional = true}
pylint = {version = "^2.17.0", optional = true}
radon = {version = "^5.1.0", optional = true}
astroid = {version = "^2.15.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
hypothesis = "^6.70.0"
sphinx = "^6.2.0"
mkdocs = "^1.4.2"
pylint = "^2.17.0"
radon = "^5.1.0"
astroid = "^2.15.0"

[tool.poetry.extras]
testing = ["pytest", "hypothesis"]
docs = ["sphinx", "mkdocs"]
analysis = ["pylint", "radon", "astroid"]

[tool.poetry.scripts]
zangalewa = "zangalewa.cli.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "invalid-name",
]

[tool.pylint.format]
max-line-length = 100 